+++ etc/env.dist
@@ -25,7 +25,7 @@ WEBSITE_STORAGE_URL=changeme
 SOURCE_BUCKET=changeme
 HOSTED_ZONE_ID=changeme
 
-# Broswerless
+# Browserless
 BROWSERLESS_ENDPOINT=dm-broswerless:3000
 MAX_CONCURRENT_SESSIONS=10
